pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cyberark-identity-library"
version = "0.1.0"
description = "A library for managing identities in CyberArk."
authors = [{ name = "Your Name", email = "your.email@example.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.6"
keywords = ["cyberark", "identity", "library"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests",
    "python-dotenv"
]

[project.urls]
Homepage = "https://github.com/gravitylens/cyberark-identity-library"
Repository = "https://github.com/gravitylens/cyberark-identity-library"